Get rid of ssize_t.
This is a wrong return type for functions that return negative value on
error and byte count on success.

Functions that take a memory buffer are never passed buffers exceeding
INT_MAX bytes in size, therefore they can return a plain int.

Functions that return file sizes or positions can return values
exceeding INT_MAX, therefore they must return int64_t.

+++ src/common/cmd.c
@@ -1575,7 +1575,7 @@ void Cmd_ExecuteString(cmdbuf_t *buf, const char *text)
 qerror_t Cmd_ExecuteFile(const char *path, unsigned flags)
 {
     char    *f;
-    ssize_t len;
+    int len;
     qerror_t ret;
     cmdbuf_t *buf;
